feat(sidenav): extract navitem story to it's own file/dir
affects: @crave/farmblocks-sidenav
